cleaner logging message
